requirements.txt
fastapi
uvicorn
opencv-python
pyzbar
python-dotenv
paho-mqtt
